refactor: divide analytics page into granular components
