+++ apps/app/components/workspace/analytics/index.ts
@@ -1,0 +1,2 @@
+export * from "./analytics-table";
+export * from "./sidebar"
