+++ apps/app/components/workspace/index.ts
@@ -1,3 +1,4 @@
+export * from "./analytics";
 export * from "./activity-graph";
 export * from "./completed-issues-graph";
 export * from "./create-workspace-form";
