+++ apps/app/constants/graph.ts
@@ -17,6 +17,7 @@ export const CHARTS_THEME: Theme = {
       background: "rgb(var(--color-bg-surface-2))",
       color: "rgb(var(--color-text-secondary))",
       fontSize: "0.8rem",
+      border: "1px solid rgb(var(--color-border))",
     },
   },
   grid: {
